Use gcc's download_prerequisites script to pull in mpfr, gmp, mpc, etc.
By pulling these in locally and compiling them as part of the toolchain build, they are built as static libraries and don't rely on shared libraries. This helps to make the buildroot more portable to another host machine. Otherwise, a downstream user must ensure they have the same version of the shared libraries.