toolchain/kconfig-frontends/Config.in:  Direct users to NuttX kconfig-frontends snapshot.
It appears Yann Morin's kconfig-frontends project has gone offline and
the website at http://ymorin.is-a-geek.org/projects/kconfig-frontends
is not responding. Until we can identify a "proper" upstream project,
use the snapshot in NuttX tools repository.

Replace (almost) all links to Yann Morin's kconfig-frontends in the
docs to point to the NuttX tools snapshot instead, with the following
exceptions:

    buildroot/toolchain/kconfig-frontends/Config.in: Document the
    source of our snapshot.